requirements.txt
ipykernel
langchain
langchain_community
langchain-core
langchain-openai
langchain-weaviate
matplotlib
numpy
pandas
python-dotenv
python-docx
scikit-learn
seaborn
torch
torchvision
transformers
weaviate-client